refactor: merge feedback tests into existing test files
Move ActionResult and flush_promises tests into test_orchestrator.py,
move fuser feedback tests into test_init.py. Reuse existing mock
infrastructure instead of duplicating it.
